pyproject.toml
[project]
name = "between"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.2",
    "django-allauth==65.9.0",
    "django-allauth[mfa]",
    "django-bootstrap5==25.1",
    "django-browser-reload>=1.18.0",
    "django-debug-toolbar==5.2.0",
    "gunicorn==23.0.0",
    "pillow==11.2.1",
    "psycopg2==2.9.10",
    "python-decouple==3.8",
    "PyYAML==6.0.2",
    "resend==2.10.0",
    "sentry-sdk==3.0.0a1",
    "whitenoise==6.9.0",
]
